[ALPS03360064] i2c: fix parsing error
[Detail]
1. fix parsing error

MTK-Commit-Id: c26969cf88fbf86cd7d07373f989144f59ed63b1

CR-Id: ALPS03360064
Feature: Others

+++ drivers/i2c/busses/i2c-mtk.h
@@ -342,10 +342,10 @@ struct mt_i2c {
 	const struct mtk_i2c_compatible *dev_comp;
 	int rec_idx; /* next record idx */
 	struct i2c_info rec_info[I2C_RECORD_LEN];
-	u16 ch_offset_default;
-	u16 ch_offset;
-	u16 dma_ch_offset_default;
-	u16 dma_ch_offset;
+	int ch_offset_default;
+	int ch_offset;
+	int dma_ch_offset_default;
+	int dma_ch_offset;
 };
 
 #if defined(CONFIG_MTK_FPGA) || defined(CONFIG_FPGA_EARLY_PORTING)
